vim: add ChangeTheme command
I was experimenting with different colorscheme lately, and changing all
the different part of the theme (vim colorscheme, vim-airline theme and
tmuxline-theme) was quite tedious.
The ChangeTheme command makes it somewhat automated.

A very cool feature to implement in the future would be to automatically
change .vimrc, airline config and .tmux-status.conf upon confirming