tmux: fix background overwriting on version >= 3.3
Version 3.3 makes it so that style option when appended don't override
the sub element (ie fg, bg), but append the additional text at the end
of the value. This broke the toggling of transparent background.
The fix here makes things a tad bit more convoluted, but make it work
for all versions of tmux