tmux: refactor tmuxline element (to make use of new tmux_mode_style)
The way that tmuxline works prevents us for having a nicely colored
element as the last element of the right of left status (it will look
bad with the way that tmuxline places style element in the string).
So similar as the status-right elements (where the mode indicator is put
in 'y', not 'z'. We place the session indicator in 'b', not 'a'.
Since we don't want to place anything in those other elements, we're
good.