tmux: Exit from copy-mode on ":"
Sometimes copy-mode starts by mistake, especially in vim when doing
mouse work, and stays on. So next time I tried to type a vim command the
pane goes haywire.
When I use ":" it's almost entirely for entering command mode in vim, so
let's just disable copy-mode and resend the key if it was on.